requirements.txt
tqdm
torch
dask
torchmetrics
pillow
torchvision
matplotlib
